moves E12AuditProgressFactory into own file
