+++ epilepsy12/tests/factories/factories.py
@@ -15,39 +15,10 @@ from epilepsy12.models import (
     Organisation,
     Site,
     Registration,
-    AuditProgress,
     KPI,
 )
 from epilepsy12.constants import user_types, VALID_NHS_NUMS
-
-
-class E12AuditProgressFactory(factory.django.DjangoModelFactory):
-    """Factory fn to create new E12 AuditProgress"""
-
-    class Meta:
-        model = AuditProgress
-
-    registration_complete = False
-    first_paediatric_assessment_complete = False
-    assessment_complete = False
-    epilepsy_context_complete = False
-    multiaxial_diagnosis_complete = False
-    management_complete = False
-    investigations_complete = False
-    registration_total_expected_fields = 3
-    registration_total_completed_fields = 0
-    first_paediatric_assessment_total_expected_fields = 0
-    first_paediatric_assessment_total_completed_fields = 0
-    assessment_total_expected_fields = 0
-    assessment_total_completed_fields = 0
-    epilepsy_context_total_expected_fields = 0
-    epilepsy_context_total_completed_fields = 0
-    multiaxial_diagnosis_total_expected_fields = 0
-    multiaxial_diagnosis_total_completed_fields = 0
-    investigations_total_expected_fields = 0
-    investigations_total_completed_fields = 0
-    management_total_expected_fields = 0
-    management_total_completed_fields = 0
+from .E12AuditProgressFactory import E12AuditProgressFactory
 
 
 class E12SiteFactory(factory.django.DjangoModelFactory):
